+++ src/main/java/org/synyx/urlaubsverwaltung/workingtime/WorkingTimeCalendarServiceImpl.java
@@ -1,0 +1,153 @@
+package org.synyx.urlaubsverwaltung.workingtime;
+
+import org.springframework.stereotype.Service;
+import org.synyx.urlaubsverwaltung.absence.DateRange;
+import org.synyx.urlaubsverwaltung.period.DayLength;
+import org.synyx.urlaubsverwaltung.person.Person;
+import org.synyx.urlaubsverwaltung.publicholiday.PublicHoliday;
+import org.synyx.urlaubsverwaltung.publicholiday.PublicHolidaysService;
+import org.synyx.urlaubsverwaltung.settings.SettingsService;
+
+import java.time.DayOfWeek;
+import java.time.LocalDate;
+import java.time.Year;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.function.Supplier;
+
+import static java.time.temporal.TemporalAdjusters.lastDayOfYear;
+import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toMap;
+
+@Service
+class WorkingTimeCalendarServiceImpl implements WorkingTimeCalendarService {
+
+    private final WorkingTimeRepository workingTimeRepository;
+    private final PublicHolidaysService publicHolidaysService;
+    private final SettingsService settingsService;
+
+    WorkingTimeCalendarServiceImpl(WorkingTimeRepository workingTimeRepository, PublicHolidaysService publicHolidaysService, SettingsService settingsService) {
+        this.workingTimeRepository = workingTimeRepository;
+        this.publicHolidaysService = publicHolidaysService;
+        this.settingsService = settingsService;
+    }
+
+    @Override
+    public Map<Person, WorkingTimeCalendar> getWorkingTimesByPersons(Collection<Person> persons, Year year) {
+        return getWorkingTimesByPersons(persons, new DateRange(year.atDay(1), year.atDay(1).with(lastDayOfYear())));
+    }
+
+    @Override
+    public Map<Person, WorkingTimeCalendar> getWorkingTimesByPersons(Collection<Person> persons, DateRange dateRange) {
+        final CachedSupplier<FederalState> federalStateCachedSupplier = new CachedSupplier<>(this::getSystemDefaultFederalState);
+
+        final WorkingTimeSettings workingTimeSettings = settingsService.getSettings().getWorkingTimeSettings();
+
+        final Map<Person, List<WorkingTime>> workingTimesByPerson = workingTimeRepository.findByPersonIsInOrderByValidFromDesc(persons)
+            .stream()
+            .map(entity -> toWorkingTime(entity, federalStateCachedSupplier))
+            .collect(groupingBy(WorkingTime::getPerson));
+
+        final LocalDate start = dateRange.getStartDate();
+        final LocalDate end = dateRange.getEndDate();
+
+        return persons.stream().map(person -> {
+            final List<WorkingTime> workingTimesInDateRange = workingTimesByPerson.getOrDefault(person, List.of())
+                .stream()
+                .filter(workingTime -> !workingTime.getValidFrom().isAfter(end))
+                .collect(toList());
+
+            final Map<LocalDate, DayLength> dayLengthByDate = new HashMap<>();
+
+            LocalDate nextEnd = end;
+
+            for (WorkingTime workingTime : workingTimesInDateRange) {
+                final FederalState federalState = workingTime.getFederalState();
+
+                final DateRange workingTimeDateRange;
+                if (workingTime.getValidFrom().isBefore(start)) {
+                    workingTimeDateRange = new DateRange(start, nextEnd);
+                } else {
+                    workingTimeDateRange = new DateRange(workingTime.getValidFrom(), nextEnd);
+                }
+
+                for (LocalDate date : workingTimeDateRange) {
+                    DayLength dayLengthForWeekDay = workingTime.getDayLengthForWeekDay(date.getDayOfWeek());
+                    if (dayLengthForWeekDay.getDuration().signum() > 0) {
+                        final Optional<PublicHoliday> maybePublicHoliday = publicHolidaysService.getPublicHoliday(date, federalState, workingTimeSettings);
+
+                        if (maybePublicHoliday.isPresent()) {
+                            final PublicHoliday publicHoliday = maybePublicHoliday.get();
+                            if (dayLengthForWeekDay.equals(DayLength.FULL)) {
+                                dayLengthForWeekDay = publicHoliday.getDayLength().getInverse();
+                            } else {
+                                if (dayLengthForWeekDay.equals(DayLength.MORNING)) {
+                                    if (publicHoliday.isFull() || publicHoliday.isMorning()) {
+                                        dayLengthForWeekDay = DayLength.ZERO;
+                                    }
+                                } else {
+                                    if (publicHoliday.isFull() || publicHoliday.isNoon()) {
+                                        dayLengthForWeekDay = DayLength.ZERO;
+                                    }
+                                }
+                            }
+                        }
+
+                    }
+                    dayLengthByDate.put(date, dayLengthForWeekDay);
+                }
+
+                if (workingTimeDateRange.getStartDate().equals(start)) {
+                    break;
+                }
+
+                nextEnd = workingTime.getValidFrom().minusDays(1);
+            }
+
+            return Map.entry(person, new WorkingTimeCalendar(dayLengthByDate));
+        }).collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
+    }
+
+    private static WorkingTime toWorkingTime(WorkingTimeEntity workingTimeEntity, Supplier<FederalState> defaultFederalStateProvider) {
+
+        final boolean isDefaultFederalState = workingTimeEntity.getFederalStateOverride() == null;
+        final FederalState federalState = isDefaultFederalState ? defaultFederalStateProvider.get() : workingTimeEntity.getFederalStateOverride();
+
+        final WorkingTime workingTime = new WorkingTime(workingTimeEntity.getPerson(), workingTimeEntity.getValidFrom(), federalState, isDefaultFederalState);
+
+        for (DayOfWeek dayOfWeek : DayOfWeek.values()) {
+            final DayLength dayLength = dayLengthForDayOfWeek(workingTimeEntity, dayOfWeek);
+            workingTime.setDayLengthForWeekDay(dayOfWeek, dayLength);
+        }
+
+        return workingTime;
+    }
+
+    private static DayLength dayLengthForDayOfWeek(WorkingTimeEntity workingTimeEntity, DayOfWeek dayOfWeek) {
+        switch (dayOfWeek) {
+            case MONDAY:
+                return workingTimeEntity.getMonday();
+            case TUESDAY:
+                return workingTimeEntity.getTuesday();
+            case WEDNESDAY:
+                return workingTimeEntity.getWednesday();
+            case THURSDAY:
+                return workingTimeEntity.getThursday();
+            case FRIDAY:
+                return workingTimeEntity.getFriday();
+            case SATURDAY:
+                return workingTimeEntity.getSaturday();
+            case SUNDAY:
+                return workingTimeEntity.getSunday();
+        }
+        return DayLength.ZERO;
+    }
+
+    private FederalState getSystemDefaultFederalState() {
+        return settingsService.getSettings().getWorkingTimeSettings().getFederalState();
+    }
+}
