+++ src/main/java/org/synyx/urlaubsverwaltung/workingtime/WorkingTimeServiceImpl.java
@@ -7,15 +7,11 @@ import org.springframework.transaction.annotation.Transactional;
 import org.synyx.urlaubsverwaltung.absence.DateRange;
 import org.synyx.urlaubsverwaltung.period.DayLength;
 import org.synyx.urlaubsverwaltung.person.Person;
-import org.synyx.urlaubsverwaltung.publicholiday.PublicHoliday;
-import org.synyx.urlaubsverwaltung.publicholiday.PublicHolidaysService;
 import org.synyx.urlaubsverwaltung.settings.SettingsService;
 
 import java.time.Clock;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
-import java.time.Year;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -25,8 +21,6 @@ import java.util.function.Supplier;
 import static java.lang.invoke.MethodHandles.lookup;
 import static java.time.format.DateTimeFormatter.ofPattern;
 import static java.time.temporal.TemporalAdjusters.firstDayOfYear;
-import static java.time.temporal.TemporalAdjusters.lastDayOfYear;
-import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 import static org.slf4j.LoggerFactory.getLogger;
@@ -40,16 +34,14 @@ class WorkingTimeServiceImpl implements WorkingTimeService, WorkingTimeWriteServ
 
     private final WorkingTimeProperties workingTimeProperties;
     private final WorkingTimeRepository workingTimeRepository;
-    private final PublicHolidaysService publicHolidaysService;
     private final SettingsService settingsService;
     private final Clock clock;
 
     @Autowired
     public WorkingTimeServiceImpl(WorkingTimeProperties workingTimeProperties, WorkingTimeRepository workingTimeRepository,
-                                  PublicHolidaysService publicHolidaysService, SettingsService settingsService, Clock clock) {
+                                  SettingsService settingsService, Clock clock) {
         this.workingTimeProperties = workingTimeProperties;
         this.workingTimeRepository = workingTimeRepository;
-        this.publicHolidaysService = publicHolidaysService;
         this.settingsService = settingsService;
         this.clock = clock;
     }
@@ -132,82 +124,6 @@ class WorkingTimeServiceImpl implements WorkingTimeService, WorkingTimeWriteServ
         }
 
         return workingTimesOfPersonByDateRange;
-    }
-
-    @Override
-    public Map<Person, WorkingTimeCalendar> getWorkingTimesByPersons(Collection<Person> persons, Year year) {
-        return getWorkingTimesByPersons(persons, new DateRange(year.atDay(1), year.atDay(1).with(lastDayOfYear())));
-    }
-
-    @Override
-    public Map<Person, WorkingTimeCalendar> getWorkingTimesByPersons(Collection<Person> persons, DateRange dateRange) {
-        final CachedSupplier<FederalState> federalStateCachedSupplier = new CachedSupplier<>(this::getSystemDefaultFederalState);
-
-        final WorkingTimeSettings workingTimeSettings = settingsService.getSettings().getWorkingTimeSettings();
-
-        final Map<Person, List<WorkingTime>> workingTimesByPerson = workingTimeRepository.findByPersonIsInOrderByValidFromDesc(persons)
-            .stream()
-            .map(entity -> toWorkingTime(entity, federalStateCachedSupplier))
-            .collect(groupingBy(WorkingTime::getPerson));
-
-        final LocalDate start = dateRange.getStartDate();
-        final LocalDate end = dateRange.getEndDate();
-
-        return persons.stream().map(person -> {
-            final List<WorkingTime> workingTimesInDateRange = workingTimesByPerson.getOrDefault(person, List.of())
-                .stream()
-                .filter(workingTime -> !workingTime.getValidFrom().isAfter(end))
-                .collect(toList());
-
-            final Map<LocalDate, DayLength> dayLengthByDate = new HashMap<>();
-
-            LocalDate nextEnd = end;
-
-            for (WorkingTime workingTime : workingTimesInDateRange) {
-                final FederalState federalState = workingTime.getFederalState();
-
-                final DateRange workingTimeDateRange;
-                if (workingTime.getValidFrom().isBefore(start)) {
-                    workingTimeDateRange = new DateRange(start, nextEnd);
-                } else {
-                    workingTimeDateRange = new DateRange(workingTime.getValidFrom(), nextEnd);
-                }
-
-                for (LocalDate date : workingTimeDateRange) {
-                    DayLength dayLengthForWeekDay = workingTime.getDayLengthForWeekDay(date.getDayOfWeek());
-                    if (dayLengthForWeekDay.getDuration().signum() > 0) {
-                        final Optional<PublicHoliday> maybePublicHoliday = publicHolidaysService.getPublicHoliday(date, federalState, workingTimeSettings);
-
-                        if (maybePublicHoliday.isPresent()) {
-                            final PublicHoliday publicHoliday = maybePublicHoliday.get();
-                            if (dayLengthForWeekDay.equals(DayLength.FULL)) {
-                                dayLengthForWeekDay = publicHoliday.getDayLength().getInverse();
-                            } else {
-                                if (dayLengthForWeekDay.equals(DayLength.MORNING)) {
-                                    if (publicHoliday.isFull() || publicHoliday.isMorning()) {
-                                        dayLengthForWeekDay = DayLength.ZERO;
-                                    }
-                                } else {
-                                    if (publicHoliday.isFull() || publicHoliday.isNoon()) {
-                                        dayLengthForWeekDay = DayLength.ZERO;
-                                    }
-                                }
-                            }
-                        }
-
-                    }
-                    dayLengthByDate.put(date, dayLengthForWeekDay);
-                }
-
-                if (workingTimeDateRange.getStartDate().equals(start)) {
-                    break;
-                }
-
-                nextEnd = workingTime.getValidFrom().minusDays(1);
-            }
-
-            return Map.entry(person, new WorkingTimeCalendar(dayLengthByDate));
-        }).collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
     }
 
     @Override
@@ -330,22 +246,5 @@ class WorkingTimeServiceImpl implements WorkingTimeService, WorkingTimeWriteServ
                 return workingTimeEntity.getSunday();
         }
         return DayLength.ZERO;
-    }
-
-    private static class CachedSupplier<T> implements Supplier<T> {
-        private T cachedValue;
-        private final Supplier<T> supplier;
-
-        CachedSupplier(Supplier<T> supplier) {
-            this.supplier = supplier;
-        }
-
-        @Override
-        public T get() {
-            if (cachedValue == null) {
-                cachedValue = supplier.get();
-            }
-            return cachedValue;
-        }
     }
 }
